Resolve "Integrate pre-transition 1.6 work-in-process github branch into gitlab develop"

Closes #83

See merge request opencpi/opencpi!6

+++ projects/assets/components/dsp_comps/iq_imbalance_fixer.test/DFTCalculator.py
@@ -172,7 +172,7 @@ class DFTCalculator():
             msg += ", supported values are None and dB_relative_to_unity"
             raise Exception(msg)
         nn = self.result.get_num_dft_points()
-        result_pos_freqs = self.result.amplitudes[0:(nn/2)-1]
+        result_pos_freqs = self.result.amplitudes[0:int(nn/2)-1]
         tmp = 20*np.log10(abs(result_pos_freqs) + eps)
         ret = tmp[np.argmax(tmp)]
         return ret
@@ -189,7 +189,7 @@ class DFTCalculator():
             msg += ", supported values are None and dB_relative_to_unity"
             raise Exception(msg)
         nn = self.result.get_num_dft_points()
-        result_neg_freqs = self.result.amplitudes[nn/2:nn-1]
+        result_neg_freqs = self.result.amplitudes[int(nn/2):nn-1]
         tmp = 20*np.log10(abs(result_neg_freqs) + eps)
         ret = tmp[np.argmax(tmp)]
         return ret
